api/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Opcional: para logging, métricas, etc.
loguru==0.7.2